refactor: move env constant to index file
