+++ src/utils/constants/index.ts
@@ -3,3 +3,5 @@ export const Env = {
   DEVELOPMENT: 'development',
   STAGING: 'staging',
 };
+
+export const IS_PUBLIC_KEY = 'isPublic';
